requirements.txt
valr-python
pandas>=2.0.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
websockets==10.4
